requirements.txt
groq>=1.0.0
nanda-adapter>=0.1.0
rich>=13.0.0
matplotlib>=3.8.0
python-dotenv>=1.0.0
requests>=2.31.0
